backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "hiredis>=3.2.1",
    "httpx>=0.28.1",
    "langchain-core>=0.3.72",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.2",
    "loguru>=0.7.3",
    "matplotlib>=3.10.5",
    "mplfinance>=0.12.10b0",
    "numpy<2.0",
    "pandas>=2.3.1",
    "pandas-ta>=0.3.14b0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "redis>=6.2.0",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.42",
    "supabase>=2.17.0",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
    "google-generativeai>=0.8.3",
    "langchain>=0.3.27",
]
